Fixed some logic error

+++ cpanel/webmaster/users$1.php
@@ -18,10 +18,10 @@ if (isset($array['search_button']) || //$array from index.php
     $order = html_entity_decode(filter_input(INPUT_POST, "sort_order"));
 
     try {
-        if (isset($array['delete_button'])) {
+        if (isset($array['delete_button']) && isset($array['checkbox'])) {
             $actionPerformed = true;
             deleteUsers($array['checkbox']);
-        } elseif (isset($array['activate_button'])) {
+        } elseif (isset($array['activate_button']) && isset($array['checkbox'])) {
             $actionPerformed = true;
             activateUsers($array['checkbox']);
         }
@@ -69,7 +69,7 @@ limitations under the License.
 <script>
     function warn() {
         var ok = confirm("Are you sure?");
-        if(ok === false){
+        if (ok === false) {
             //Cancel request
             window.stop();
         }
@@ -86,7 +86,7 @@ limitations under the License.
     </div>
     <div class="row">
         <?php
-        if (empty($users) and !isset($array['search_button']) ) {
+        if (empty($users) and ! isset($array['search_button'])) {
             echo '<p>No suspended user</p>';
         } else {
             ?>
@@ -189,7 +189,7 @@ limitations under the License.
                                         <td class="text-left"><?= $users[$index]['department'] ?></td>
                                         <td class="text-left"><?= $users[$index]['level'] ?></td>
                                         <td class="text-left">
-                                            <a href="index.php?p=13&url=<?= urlencode(CPANEL_URL . 'webmaster/?p=11&pg=' . $page) ?>&id=<?= $users[$index]['regno'] ?>&search=<?= urlencode($searchQuery) ?>&sort_type=<?= $sort_type ?>&sort_order=<?= $order ?>">
+                                            <a href="index.php?p=13&url=<?= urlencode(CPANEL_URL . 'webmaster/?p=11&pg=' . $page . '&q=' . urlencode($searchQuery) . '&s=' . $sort_type . '&o=' . $order) ?>&id=<?= $users[$index]['regno'] ?>">
                                                 preview
                                             </a>
                                         </td>
